requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
greenlet==3.3.1
bcrypt==4.1.3
alembic==1.13.1
PyJWT==2.11.0
